Create a simple controller with tests
The controller echoes HTTP headers as a JSON.  The headers are
transformed into environment variables in all caps with `_` between
words and prefixed with `HTTP_`.  The controller removes the prefix but
leaves them otherwise as is.

The headers in the controller look somewhat like a hash, but they are
a separate class with some methods that make it _look_ like a hash,
but it is much more limited.

Ran `rake db:create` and `rake db:migrate` so the tests would stop
complaining about the missing database schema.